Fixes logic error leading to unuseful error
When opening/creating an Index, the code currently assumes that
`Index::open_in_dir` only returns errors for `meta.json` not
existing, when that is not the case. This can lead to the code
trying to create a new index when there already is one present.

This commit just makes it give a more informative panic in the
case that it gives a different error.

See issue #54